move private function to utils.go
